third_party/WebKit/LayoutTests/bluetooth/requestDevice/canonicalizeFilter/unicode-max-length-for-name-in-adv-name.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script>
promise_test(() => {
  // \u2764's UTF-8 respresentation is 3 bytes long.
  // 83 chars * 3 bytes/char = 249 bytes
  let unicode_name = generate_string(83, '\u2764');
  return assert_promise_rejects_with_message(
    requestDeviceWithKeyDown({filters: [{name: unicode_name}]}),
    new DOMException(
      'Failed to execute \'requestDevice\' on \'Bluetooth\': ' +
      'A device name can\'t be longer than 248 bytes.',
    'TypeError'));
}, 'Unicode string in \'name\' with utf8 representation greater than 248 ' +
   'bytes must throw TypeError.');
</script>
